requirements.txt
streamlit
pandas
numpy
scikit-learn
sentence-transformers
torch
python-docx
PyPDF2
